Retain trailing newline.
During #3, there was some debate about how newlines should be preserved,
especially with regards to the trailing newline at the end of the
program's output.

While the trailing newline may make the test expectations a little more
ugly, my feeling is it's more correct when compared to what the `jq`
binary itself produces. A difficult decision to make at this point is
whether or not this diff should be considered a breaking change, bumping
us up to 0.4.x or if it should be a plain bugfix, continuing the current
0.3.x line.

Best as I can tell, callers who are relying on feeding the output into a
parser such as serde will not see any impact. The only callers who may
have been impacted are those who are doing string matching on the raw
string. I'm going to sleep on the matter of which digit to bump the
version by. I may use this as an opportinity to make some other changes
with the opening this creates, perhaps in error handling.