Implements std::error::Error for jq_rs::errors::Error.
Going with a conservative approach for errors to start with. This diff
adds an impl for the stdlib Error trait, and ensures there are From
impls for any other error types which can surface through this lib.

This also provides a custom Result type, which is adopted throughout.

Added the `matches` crate as a dev dep to make the tests which look at
the error types easier to write.