Fixes jv_copy/jv_free symmetry.
Fixes the existing fixmes around what should have been a redundant
jv_copy and a seemingly missing jv_free.

When the copies and frees were in the correct spots, we could see an
`invalid pointer` error. I'm assuming this is a double-free.

I don't know the implementation details on the libjq side, but the
double-free seems to be resolved by treating `jq_start()` as a consuming
call. In `jq` I don't see that initial value being freed explicitly, so
this is likely a good bet. Since our initial value is wrapped in a `JV`,
we can't avoid the drop, so as a workaround I've added a jv_copy to that
`jq_start` call.