Remove Error::cause impl since it is deprecated.
Looks like `Error::source` replaces `Error::cause`. Originally I added
them both since I wasn't sure it made sense to only include one or the
other, but since our lowest supported rust version is 1.32, and `source`
was added in 1.30, and `cause` was deprecated as of 1.33, anyone using
this crate will for sure be able to work with `source`.